+++ salalib/segmmodules/segmangular.cpp
@@ -101,7 +101,7 @@ bool SegmentAngular::run(Communicator * comm, const Options &options, ShapeGraph
                     for (auto &segconn : line.m_forward_segconns) {
                         if (!covered[segconn.first.ref]) {
                             double angle = depth_to_line + segconn.second;
-                            int rbin = lineindex.coverage;
+                            size_t rbin = lineindex.coverage;
                             while (rbin != radii.size() && radii[rbin] != -1 && angle > radii[rbin]) {
                                 rbin++;
                             }
@@ -117,7 +117,7 @@ bool SegmentAngular::run(Communicator * comm, const Options &options, ShapeGraph
                     for (auto &segconn : line.m_back_segconns) {
                         if (!covered[segconn.first.ref]) {
                             double angle = depth_to_line + segconn.second;
-                            int rbin = lineindex.coverage;
+                            size_t rbin = lineindex.coverage;
                             while (rbin != radii.size() && radii[rbin] != -1 && angle > radii[rbin]) {
                                 rbin++;
                             }
